Add trivial setup
