+++ backend-ep/README.md
@@ -1,15 +1,3 @@
-# backend-ep
+# Perform Prisma migration inside BE container
 
-To install dependencies:
-
-```bash
-bun install
-```
-
-To run:
-
-```bash
-bun run index.ts
-```
-
-This project was created using `bun init` in bun v1.2.2. [Bun](https://bun.sh) is a fast all-in-one JavaScript runtime.
+`docker-compose exec backend bunx prisma migrate dev --name init`
