+++ backend-ep/package.json
@@ -3,7 +3,11 @@
   "module": "index.ts",
   "type": "module",
   "scripts": {
-    "dev": "bun --hot src/index.ts"
+    "dev": "bun --hot src/index.ts",
+    "build": "bunx tsc",
+    "start": "node build/index.js",
+    "prisma:generate": "bunx prisma generate",
+    "prisma:migrate": "bunx prisma migrate deploy"
   },
   "devDependencies": {
     "@types/bun": "latest",
