+++ backend-ep/src/routes/user.route.ts
@@ -1,0 +1,17 @@
+import { Router } from 'express';
+import { prisma } from '../lib/prisma';
+
+const router = Router();
+
+router.get('/', async (req, res) => {
+    const users = await prisma.user.findMany();
+    res.json(users);
+});
+
+router.post('/', async (req, res) => {
+    const { name, email } = req.body;
+    const user = await prisma.user.create({ data: { name, email } });
+    res.status(201).json(user);
+});
+
+export default router;
